fastapi-backend/nutritionapi/requirements.txt
azure-functions
fastapi
PyMySQL
python-dotenv
uvicorn
python-jose


